Apples Can Spawn at Bottom/Far Right
Before the random number for x could only go to 6 because if it went to 7 and 50 was added to that, then it would spawn the apple off screen. Similar situation for the y position. Now that has been accounted for and they can spawn at the bottom row and the far right column.